Added SEO (#18)
As a total last ditch effort to enable social media previews and for better SEO, I have opted to use next-seo in the application.

This commit removes the previous implementation of generating <meta /> tags using getStaticProps() in the server and delegates the full job of SEO to the node module. Not sure whether it will work though.